Added quiz version to retain the history of quiz modification

+++ Quiz service/Quiz-Service/src/main/java/com/dhanashri/Quiz_Service/Module/ManualQuizRequest.java
@@ -10,7 +10,7 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ManualQuizRequest {
-    private Integer quizId;
+    private Long quizId;
     private String quizTitle;
     private List<ManualQuizDTO> configList;
     private String mode;//
